requirements.txt
allure-pytest==2.12.0
pytest==7.2.1
selenium==4.8.0
webdriver-manager==3.8.5
